feat: extract ReportTemplateCard component for better modularity and maintainability
